+++ test/framework/src/main/java/org/opensearch/test/InternalTestCluster.java
@@ -2625,6 +2625,7 @@ public final class InternalTestCluster extends TestCluster {
                     false,
                     false,
                     false,
+                    false,
                     false
                 );
                 assertThat(